+++ lib/bluedroid/src/btc/core/btc_storage.c
@@ -338,7 +338,7 @@ bt_status_t btc_storage_get_bonded_bt_devices_list(bt_bdaddr_t *bond_dev, int *d
     return BT_STATUS_SUCCESS;
 }
 
-bt_status_t btc_storage_get_bonded_bt_device_property(bt_bdaddr_t bd_addr, const char *key, char *value, int *len)
+bt_status_t btc_storage_get_bonded_bt_device_property(bt_bdaddr_t bd_addr, const char *key, char *value, size_t *len)
 {
     bdstr_t bdstr;
     bdaddr_to_string(&bd_addr, bdstr, sizeof(bdstr));
@@ -350,7 +350,7 @@ bt_status_t btc_storage_get_bonded_bt_device_property(bt_bdaddr_t bd_addr, const
     return ret ? BT_STATUS_SUCCESS : BT_STATUS_FAIL;
 }
 
-bt_status_t btc_storage_set_bonded_bt_device_property(bt_bdaddr_t bd_addr, const char *key, const char *value, int len)
+bt_status_t btc_storage_set_bonded_bt_device_property(bt_bdaddr_t bd_addr, const char *key, const char *value, size_t len)
 {
     bdstr_t bdstr;
     bdaddr_to_string(&bd_addr, bdstr, sizeof(bdstr));
